fix : [#29] 기능 변경 및 테이블 구조 변경으로 인한 엔티티 변경
- 기존 검색 기능 방식에서 최근 상영중인 영화들을 기반으로 서비스 방식 변경
- 엔티티 변경
- 테이블 삭제 및 수정
