+++ src/main/java/com/hackathonteam2/recomovie/cinema/dto/CinemaDto.java
@@ -1,8 +1,13 @@
 package com.hackathonteam2.recomovie.cinema.dto;
 
+import com.hackathonteam2.recomovie.cinema.entity.Cinema;
 
-import com.hackathonteam2.recomovie.cinema.Cinema;
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.ToString;
 
 @Getter
 @Setter
@@ -12,25 +17,25 @@ import lombok.*;
 @ToString
 public class CinemaDto {
 
-    private String brand;
+	private String brand;
 
-    private String region;
+	private String region;
 
-    private String name;
+	private String name;
 
-    public Cinema toEntity() {
-        return Cinema.builder()
-                .brand(this.brand)
-                .region(this.region)
-                .name(this.name)
-                .build();
-    }
+	public Cinema toEntity() {
+		return Cinema.builder()
+			.brand(this.brand)
+			.region(this.region)
+			.name(this.name)
+			.build();
+	}
 
-    public static CinemaDto of(Cinema cinema) {
-        return CinemaDto.builder()
-                .brand(cinema.getBrand())
-                .region(cinema.getRegion())
-                .name(cinema.getName())
-                .build();
-    }
+	public static CinemaDto of(Cinema cinema) {
+		return CinemaDto.builder()
+			.brand(cinema.getBrand())
+			.region(cinema.getRegion())
+			.name(cinema.getName())
+			.build();
+	}
 }
