+++ src/main/java/com/hackathonteam2/recomovie/movie/dto/MovieResponseDto.java
@@ -1,10 +1,15 @@
 package com.hackathonteam2.recomovie.movie.dto;
 
-import com.hackathonteam2.recomovie.movie.entity.Movie;
-import lombok.*;
-import org.springframework.beans.factory.annotation.Value;
-
 import java.util.List;
+
+import com.hackathonteam2.recomovie.movie.entity.Movie;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.ToString;
 
 @Getter
 @Setter
@@ -14,26 +19,24 @@ import java.util.List;
 @ToString
 public class MovieResponseDto {
 
-    private final static String imageUrl = "https://image.tmdb.org/t/p/original/";
+	private final static String imageUrl = "https://image.tmdb.org/t/p/original/";
 
-    private Long movieId;
-    private String title;
-    private String overview;
-    private String releaseDate;
-    private String poster;
-    private List<String> genres;
-    private Double popularity;
-    public static MovieResponseDto of(Movie movie) {
-        return MovieResponseDto.builder()
-                .movieId(movie.getMovieId())
-                .title(movie.getTitle())
-                .overview(movie.getOverview())
-                .releaseDate(movie.getReleaseDate())
-                .poster(imageUrl+movie.getPoster())
-                .genres(movie.getGenres().stream()
-                        .map(g->g.getGenre().getName())
-                        .toList())
-                .popularity(movie.getPopularity())
-                .build();
-    }
+	private Long movieId;
+	private String title;
+	private String overview;
+	private String releaseDate;
+	private String poster;
+	private List<String> genres;
+	private Double popularity;
+
+	public static MovieResponseDto of(Movie movie) {
+		return MovieResponseDto.builder()
+			.movieId(movie.getMovieId())
+			.title(movie.getTitle())
+			.overview(movie.getOverview())
+			.releaseDate(movie.getReleaseDate())
+			.poster(imageUrl + movie.getPoster())
+			.popularity(movie.getPopularity())
+			.build();
+	}
 }
