--- src/main/java/com/hackathonteam2/recomovie/movie/entity/MovieGenre.java
@@ -1,25 +1,0 @@
-package com.hackathonteam2.recomovie.movie.entity;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Entity
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class MovieGenre {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne
-    @JoinColumn(name = "movie_id")
-    private Movie movie;
-
-    @ManyToOne
-    @JoinColumn(name = "genre_id")
-    private Genre genre;
-}
