--- src/main/java/com/hackathonteam2/recomovie/movie/repository/GenreRepository.java
@@ -1,15 +1,0 @@
-package com.hackathonteam2.recomovie.movie.repository;
-
-import com.hackathonteam2.recomovie.movie.entity.Genre;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import javax.swing.text.html.Option;
-import java.util.Optional;
-
-@Repository
-public interface GenreRepository extends JpaRepository<Genre,Long> {
-    Genre findByGenreId(Long genreId);
-
-    Optional<Genre> findByGenreIdAndName(Long genreId, String name);
-}
